Wire mesh control pack into ExCtrl
Instantiate and clock `ExCtrlMeshCntlPack` from `ExCtrl`, then connect its inputs to the existing decoder, row-address, feed-signal, and tag-select outputs so mesh-control packet generation is available inside the execute controller. The change also adds/reset the new pre-mul and im2col state signals and narrows `c_rows`/`c_cols` from `u32` to `u16` in the pack interface and testbench to match the expected command field width.

+++ smesh/include/ExCtrlMeshCntlPack.hpp
@@ -48,8 +48,8 @@ class ExCtrlMeshCntlPack : public Component {
   Input(u32, b_unpadded_cols);
   Input(u32, d_unpadded_cols);
   Input(SmeshLocalAddr, c_addr);
-  Input(u32, c_rows);
-  Input(u32, c_cols);
+  Input(u16, c_rows);
+  Input(u16, c_cols);
   Input(bit, a_transpose);
   Input(bit, bd_transpose);
   Input(u32, total_rows);
